Criação de um novo DTO

+++ src/main/java/com/ibm/projetoibm/model/DTO/AprovadosDTO.java
@@ -1,5 +1,5 @@
 package com.ibm.projetoibm.model.DTO;
 
-public record AprovadosDTO(String nome, Long id) {
+public record AprovadosDTO(String nome, int id) {
 
 }
